+++ hawkbit-mgmt/hawkbit-mgmt-resource/pom.xml
@@ -99,4 +99,21 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
+
+    <!-- TODO - to be removed with hawkbit-mgmt-resource-deprecated -->
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>test-jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
 </project>
