+++ hawkbit-mgmt/hawkbit-mgmt-resource/src/main/java/org/eclipse/hawkbit/mgmt/rest/resource/MgmtTargetTagResource.java
@@ -12,14 +12,11 @@ package org.eclipse.hawkbit.mgmt.rest.resource;
 import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.stream.Collectors;
 
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.hawkbit.mgmt.json.model.PagedList;
-import org.eclipse.hawkbit.mgmt.json.model.tag.MgmtAssignedTargetRequestBody;
 import org.eclipse.hawkbit.mgmt.json.model.tag.MgmtTag;
 import org.eclipse.hawkbit.mgmt.json.model.tag.MgmtTagRequestBodyPut;
-import org.eclipse.hawkbit.mgmt.json.model.tag.MgmtTargetTagAssigmentResult;
 import org.eclipse.hawkbit.mgmt.json.model.target.MgmtTarget;
 import org.eclipse.hawkbit.mgmt.rest.api.MgmtTargetTagRestApi;
 import org.eclipse.hawkbit.mgmt.rest.resource.util.PagingUtility;
@@ -31,7 +28,6 @@ import org.eclipse.hawkbit.repository.TenantConfigurationManagement;
 import org.eclipse.hawkbit.repository.exception.EntityNotFoundException;
 import org.eclipse.hawkbit.repository.model.Target;
 import org.eclipse.hawkbit.repository.model.TargetTag;
-import org.eclipse.hawkbit.repository.model.TargetTagAssignmentResult;
 import org.eclipse.hawkbit.security.SystemSecurityContext;
 import org.eclipse.hawkbit.utils.TenantConfigHelper;
 import org.springframework.data.domain.Page;
@@ -55,12 +51,12 @@ public class MgmtTargetTagResource implements MgmtTargetTagRestApi {
 
     MgmtTargetTagResource(
             final TargetTagManagement tagManagement, final TargetManagement targetManagement,
-            final EntityFactory entityFactory, final SystemSecurityContext securityContext,
-            final TenantConfigurationManagement configurationManagement) {
+            final EntityFactory entityFactory,
+            final SystemSecurityContext securityContext, final TenantConfigurationManagement configurationManagement) {
         this.tagManagement = tagManagement;
         this.targetManagement = targetManagement;
         this.entityFactory = entityFactory;
-        this.tenantConfigHelper = TenantConfigHelper.usingContext(securityContext, configurationManagement);
+        tenantConfigHelper = TenantConfigHelper.usingContext(securityContext, configurationManagement);
     }
 
     @Override
@@ -174,7 +170,7 @@ public class MgmtTargetTagResource implements MgmtTargetTagRestApi {
     @Override
     public ResponseEntity<Void> unassignTarget(final Long targetTagId, final String controllerId) {
         log.debug("Unassign target {} for target tag {}", controllerId, targetTagId);
-        this.targetManagement.unassignTag(controllerId, targetTagId);
+        this.targetManagement.unassignTag(List.of(controllerId), targetTagId);
         return ResponseEntity.ok().build();
     }
 
@@ -197,40 +193,8 @@ public class MgmtTargetTagResource implements MgmtTargetTagRestApi {
         return ResponseEntity.ok().build();
     }
 
-    @Override
-    public ResponseEntity<MgmtTargetTagAssigmentResult> toggleTagAssignment(
-            final Long targetTagId, final List<MgmtAssignedTargetRequestBody> assignedTargetRequestBodies) {
-        log.debug("Toggle Target assignment {} for target tag {}", assignedTargetRequestBodies.size(), targetTagId);
-
-        final TargetTag targetTag = findTargetTagById(targetTagId);
-        final TargetTagAssignmentResult assigmentResult = this.targetManagement
-                .toggleTagAssignment(findTargetControllerIds(assignedTargetRequestBodies), targetTag.getName());
-
-        final MgmtTargetTagAssigmentResult tagAssigmentResultRest = new MgmtTargetTagAssigmentResult();
-        tagAssigmentResultRest.setAssignedTargets(
-                MgmtTargetMapper.toResponse(assigmentResult.getAssignedEntity(), tenantConfigHelper));
-        tagAssigmentResultRest.setUnassignedTargets(
-                MgmtTargetMapper.toResponse(assigmentResult.getUnassignedEntity(), tenantConfigHelper));
-        return ResponseEntity.ok(tagAssigmentResultRest);
-    }
-
-    @Override
-    public ResponseEntity<List<MgmtTarget>> assignTargetsByRequestBody(
-            final Long targetTagId, final List<MgmtAssignedTargetRequestBody> assignedTargetRequestBodies) {
-        log.debug("Assign targets {} for target tag {}", assignedTargetRequestBodies, targetTagId);
-        final List<Target> assignedTarget = this.targetManagement
-                .assignTag(findTargetControllerIds(assignedTargetRequestBodies), targetTagId);
-        return ResponseEntity.ok(MgmtTargetMapper.toResponse(assignedTarget, tenantConfigHelper));
-    }
-
     private TargetTag findTargetTagById(final Long targetTagId) {
         return tagManagement.get(targetTagId)
                 .orElseThrow(() -> new EntityNotFoundException(TargetTag.class, targetTagId));
-    }
-
-    private List<String> findTargetControllerIds(
-            final List<MgmtAssignedTargetRequestBody> assignedTargetRequestBodies) {
-        return assignedTargetRequestBodies.stream().map(MgmtAssignedTargetRequestBody::getControllerId)
-                .collect(Collectors.toList());
     }
 }
