+++ hawkbit-mgmt/hawkbit-mgmt-resource/src/test/java/org/eclipse/hawkbit/mgmt/rest/resource/MgmtTargetTagResourceTest.java
@@ -616,86 +616,4 @@ public class MgmtTargetTagResourceTest extends AbstractManagementApiIntegrationT
                 .andExpect(status().isOk());
         assertThat(targetManagement.findByTag(PAGE, tag.getId()).getContent()).isEmpty();
     }
-
-    // DEPRECATED scenarios
-    @Test
-    @Description("Verifes that tag assignments done through toggle API command are correctly assigned or unassigned.")
-    @ExpectEvents({
-            @Expect(type = TargetTagCreatedEvent.class, count = 1),
-            @Expect(type = TargetCreatedEvent.class, count = 2), 
-            @Expect(type = TargetUpdatedEvent.class, count = 4) })
-    public void toggleTagAssignment() throws Exception {
-        final TargetTag tag = testdataFactory.createTargetTags(1, "").get(0);
-        final int targetsAssigned = 2;
-        final List<Target> targets = testdataFactory.createTargets(targetsAssigned);
-
-        ResultActions result = toggle(tag, targets);
-
-        List<Target> updated = targetManagement.findByTag(PAGE, tag.getId()).getContent();
-
-        assertThat(updated.stream().map(Target::getControllerId).collect(Collectors.toList()))
-                .containsAll(targets.stream().map(Target::getControllerId).collect(Collectors.toList()));
-
-        result.andExpect(applyTargetEntityMatcherOnArrayResult(updated.get(0), "assignedTargets"))
-                .andExpect(applyTargetEntityMatcherOnArrayResult(updated.get(1), "assignedTargets"));
-
-        result = toggle(tag, targets);
-
-        updated = targetManagement.findAll(PAGE).getContent();
-
-        result.andExpect(applyTargetEntityMatcherOnArrayResult(updated.get(0), "unassignedTargets"))
-                .andExpect(applyTargetEntityMatcherOnArrayResult(updated.get(1), "unassignedTargets"));
-
-        assertThat(targetManagement.findByTag(PAGE, tag.getId())).isEmpty();
-    }
-
-    @Test
-    @Description("Verifies that tag assignments done through tag API command are correctly stored in the repository.")
-    @ExpectEvents({
-            @Expect(type = TargetTagCreatedEvent.class, count = 1),
-            @Expect(type = TargetCreatedEvent.class, count = 2), 
-            @Expect(type = TargetUpdatedEvent.class, count = 2) })
-    public void assignTargetsByRequestBody() throws Exception {
-        final TargetTag tag = testdataFactory.createTargetTags(1, "").get(0);
-        final int targetsAssigned = 2;
-        final List<Target> targets = testdataFactory.createTargets(targetsAssigned);
-
-        final ResultActions result = mvc
-                .perform(post(MgmtRestConstants.TARGET_TAG_V1_REQUEST_MAPPING + "/" + tag.getId() + "/assigned")
-                        .content(controllerIdsOld(
-                                targets.stream().map(Target::getControllerId).collect(Collectors.toList())))
-                        .contentType(MediaType.APPLICATION_JSON).accept(MediaType.APPLICATION_JSON))
-                .andDo(MockMvcResultPrinter.print())
-                .andExpect(status().isOk())
-                .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE));
-
-        final List<Target> updated = targetManagement.findByTag(PAGE, tag.getId()).getContent();
-
-        assertThat(updated.stream().map(Target::getControllerId).collect(Collectors.toList()))
-                .containsAll(targets.stream().map(Target::getControllerId).collect(Collectors.toList()));
-
-        result.andExpect(applyTargetEntityMatcherOnArrayResult(updated.get(0)))
-                .andExpect(applyTargetEntityMatcherOnArrayResult(updated.get(1)));
-    }
-
-    private static String controllerIdsOld(final Collection<String> ids) throws JSONException {
-        final JSONArray list = new JSONArray();
-        for (final String smID : ids) {
-            list.put(new JSONObject().put("controllerId", smID));
-        }
-
-        return list.toString();
-    }
-
-    private ResultActions toggle(final TargetTag tag, final List<Target> targets) throws Exception {
-        return mvc
-                .perform(post(MgmtRestConstants.TARGET_TAG_V1_REQUEST_MAPPING + "/" + tag.getId()
-                        + "/assigned/toggleTagAssignment")
-                        .content(controllerIdsOld(
-                                targets.stream().map(Target::getControllerId).collect(Collectors.toList())))
-                        .contentType(MediaType.APPLICATION_JSON).accept(MediaType.APPLICATION_JSON))
-                .andDo(MockMvcResultPrinter.print())
-                .andExpect(status().isOk())
-                .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE));
-    }
 }
